efl_ui_position_manager_list: set visibilitly more often
it is possible that the object returned by the batching call is NULL.
That is valid, however, if the batching call for the visibility is
getting the NULL element, but the placement is getting the none NULL
element, then the item on screen will not be visible. This commit
ensures that those items will be visible.

Co-authored-by: Mike Blumenkrantz <zmike@samsung.com>