docs: enable compression for tar file and avoid file name clash
Summary:
While the target name implies a compression we only ever put this all
into a tar archive so far. Using xz compression as we do with our source
archives.
Also adjust the file naming to avoid clashing with the source tarballs.
We used -doc with autotools before, so this is really more of a catchup.

Reviewers: bu5hm4n, segfaultxavi

Reviewed By: segfaultxavi

Subscribers: ProhtMeyhet, cedric, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D10266