mono-docs: Merge Property and getter/setter docs
Summary:
Property docs, when present, should always be included in Property and Accessor
(setter and getter) docs.
Accessor docs, when present, should be included in their accessor method, and
also in the property method, as a side note.
This patch does just that.

Depends on D10285
Ref T8309

Test Plan: Generate docs and take a look at `Efl.Ui.Calendar`, for example, which has a bit of everything.

Reviewers: lauromoura

Reviewed By: lauromoura

Subscribers: cedric, #reviewers, #committers

Tags: #efl

Maniphest Tasks: T8309

Differential Revision: https://phab.enlightenment.org/D10298